More db stuff. Installed python 3.6 venv

+++ modules/lidpix_db.py
@@ -1,7 +1,10 @@
 #!/usr/bin/env python
 
-# Here are functions for adding/removing users to the lidpix user database.
-# See show_usage() below for how to use it.
+# Lidpix database tool
+#
+# Handles the user table, gallery tables with images, 
+# and the gallery index table which lists all galleries
+
 
 
 import sqlite3, time, sys, argparse, os, subprocess
@@ -9,7 +12,6 @@ import bcrypt
 import folder
 #from folder import get_image_info
 #from users import UserDB
-
 
 
 def safe(s):
@@ -46,12 +48,12 @@ def new_table(table, db_file, schema_filename):
     try:
         conn, c = connect_to_db(db_file)
         with open(schema_filename, mode='r') as f:
-            scriptlines = "CREATE TABLE IF NOT EXISTS " + table + "\n(" + f.read() + ");"
+            scriptlines = "CREATE TABLE IF NOT EXISTS " + safe(table) + "\n(" + f.read() + ");"
         c.executescript(scriptlines)
         conn.commit()
         conn.close()
     except Exception as e:
-        print "Error when trying to create table " + table + " in db file " + db_file
+        print "Error when trying to create table " + table + " in" + db_file
         return False
     else:
         return True
@@ -74,26 +76,6 @@ def find_table(table, db_file):
         return False
     else:
         return fetched
-        
-
-#def find_row(column_to_search, value_to_match, table, db_file):
-#    
-#    """ Find first row where column_to_search has value_to_match in table.
-#    Return  """
-#    
-#    try:
-#        conn, c = connect_to_db(db_file)
-#        v = c.execute('DELETE FROM {t} WHERE {cs}="{vm}"'.format(t=safe(table),
-#                  cs=safe(column_to_search), vm=value_to_match))
-#        deleted = c.rowcount
-#        conn.commit()
-#        conn.close()
-#    except Exception as e:
-#        print "Error when trying to delete row in table", table, "in db file", db_file
-#        print e
-#        return 0
-#    else:
-#        return deleted
         
 
 def delete_table(table, db_file):
@@ -125,7 +107,7 @@ def delete_row(column_to_search, value_to_match, table, db_file):
         conn.commit()
         conn.close()
     except Exception as e:
-        print "Error when trying to delete row in table", table, "in db file", db_file
+        print "Error when trying to delete row in table", table, "in", db_file
         print e
         return 0
     else:
@@ -163,7 +145,7 @@ def get_row(column_to_search, value_to_match, table, db_file):
         conn.close()
         return row
     except Exception as e:
-        print "Error when trying to fetch row in table", table, "in database file", db_file
+        print "Error when trying to get row in table", table, "in", db_file
         print e
         return None
 
@@ -179,12 +161,12 @@ def get_all_rows(table, db_file):
         conn.close()
         return allrows
     except Exception as e:
-        print "Error when trying to fetch all rows in table", table, "in database file", db_file
+        print "Error when trying to fetch all rows in table", table, "in", db_file
         print e
         return []
         
 
-def add_gallery(galleryname, defpath, description, tags, zipfile, 
+def add_gallery(galleryname, gpath, description, tags, zipfile, 
                 users_r, users_w, groups_r, groups_w, table, db_file):
     
     """ Add a row with gallery properties to a gallery index table """
@@ -192,17 +174,17 @@ def add_gallery(galleryname, defpath, description, tags, zipfile,
     try:
         conn, c = connect_to_db(db_file)
         sql_cmd = """INSERT INTO {t}
-        (gallery_id, gallery_name, defpath, description, tags, 
+        (id, gallery_name, gpath, description, tags, 
         time_added, zipfile, users_r, users_w, groups_r, groups_w)
         VALUES (NULL, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);""".format(t=safe(table))
-        c.execute(sql_cmd, (galleryname, defpath, description, tags,
+        c.execute(sql_cmd, (galleryname, gpath, description, tags,
                             time.asctime(time.localtime(time.time())),
                             zipfile, users_r, users_w, groups_r, groups_w,))
         conn.commit()
         conn.close()
     except Exception as e:
         print "Error when trying to add gallery in table " + table + \
-        " in db file " + db_file
+        " in " + db_file
         print e
         return False
     else:
@@ -217,7 +199,7 @@ def add_images(imagefiles, description, tags, users_r,
     try:
         conn, c = connect_to_db(db_file)
         sql_cmd = """INSERT INTO {t} 
-            (image_id, imagefile, description, tags, time_photo, 
+            (id, imagefile, description, tags, time_photo, 
             time_added, users_r, users_w, groups_r, groups_w)
             VALUES (NULL, ?, ?, ?, ?, ?, ?, ?, ?, ?);""".format(t=safe(table))
         time_added = time.asctime(time.localtime(time.time()))
@@ -232,45 +214,44 @@ def add_images(imagefiles, description, tags, users_r,
         conn.commit()
         conn.close()
     except Exception as e:
-        print "Error when trying to add images in table", table, "in db file", db_file
+        print "Error when trying to add images in table", table, "in", db_file
         print e
         return False
     else:
         return True
 
     
-# Improve this - remove format, etc.
 def add_user(username, password, fullname, joined, groups, table, db_file):
     
     """ Add a lidpix user to the users table """
     
     try:
         conn, c = connect_to_db(db_file)
-        format_str = """INSERT INTO {table} (user_nr, username, password,
+        sqlcmd = """INSERT INTO {t} (id, username, password,
         fullname, groups, joining, active, confirmdelete, viewmode, theme)
-        VALUES (NULL, "{username}", "{password}", "{fullname}", "{groups}", "{joined}", 1, 1, 10, "default");"""
-        sql_command = format_str.format(table=table, username=username, password=password, \
-        fullname=fullname, joined=joined, groups=groups)
-        c.execute(sql_command)
+        VALUES (NULL, ?, ?, ?, ?, ?, 1, 1, 10, "default");""".format(t=safe(table))
+        #sql_command = format_str.format(table=safe(table), username=username, password=password, \
+        #fullname=fullname, joined=joined, groups=groups)
+        print type(password)
+        c.execute(sqlcmd, (username, password, fullname, joined, groups,))
         conn.commit()
         conn.close()
     except Exception as e:
-        print "Error when trying to add user " + username + \
-        " in table " + table + " in db file " + db_file
+        print "add_user(): Error when trying to add user " + username + \
+        " in table " + table + " in " + db_file
         print e
         return False
     else:
         return True
         
 
-# Improve this one too
 def delete_user(username, table, db_file):
     
     """ Remove a lidpix user from the users table """
     
     try:
         conn, c = connect_to_db(db_file)
-        command = "DELETE FROM {table} WHERE username =?".format(table=table)
+        command = "DELETE FROM {table} WHERE username =?".format(table=safe(table))
         c.execute(command, (username,))
         conn.commit()
         conn.close()
@@ -305,26 +286,6 @@ def print_table(table, db_file):
         print e
         
 
-# Delete this function and/or make something more generally explaining...
-# Maybe readme/doc is enough
-def show_usage():
-    print """Usage: lidpix_db.py command [username|galleryname] [-t table] [-d database_file] [images]
-          command can be:
-          addimages    add [images] to image gallery table with [galleryname]
-          addgallery   add gallery with [galleryname] to gallery index table in database_file
-          adduser      add user with [username] to user table in database_file
-          deleteuser   delete user from table in database_file
-          newutable    create new table (overwrite existing)
-          newgtable    create new table of galleries (overwrite existing)
-          printtable   print existing table in database_file
-          
-          Options:
-          --table -t   specify which table (with galleries, users) to use
-          --dbfile -d  specify which sqlite database file to use
-          """
-    sys.exit(0)
-    
-
 def get_user_input(genre, galleryname):
     
     """ Get input from user and return as a tuple, to use when adding 
@@ -348,11 +309,11 @@ def get_user_input(genre, galleryname):
     return (desc, tags, zipfile, users_r, users_w, group_r, group_w)
 
 
-def get_defpath(defpath):
+def get_gpath(gpath):
     """ Let user specify default path for images in gallery """
-    userpath = raw_input("Base path for images [%s]: " % defpath)
+    userpath = raw_input("Base path for images [%s]: " % gpath)
     if userpath == '':
-        userpath = defpath
+        userpath = gpath
     return userpath
     
 
@@ -372,10 +333,10 @@ def add_to_zip(zipfile, zippath, files):
     return True
 
 
-def new_gallery(galleryname, defpath, desc, tags, zipfile, ur, uw, gr, gw, dbfile):
+def create_gallery(galleryname, gpath, desc, tags, zipfile, ur, uw, gr, gw, dbfile):
     """ Create a new gallery """
     new_table('galleryindex', dbfile, 'gallery_db_schema.sql') # Create gallery index table if not existing
-    if add_gallery(galleryname, defpath, desc, tags, zipfile, ur, uw, gr, gw, 'galleryindex', dbfile): # Add gallery to gallery index table
+    if add_gallery(galleryname, gpath, desc, tags, zipfile, ur, uw, gr, gw, 'galleryindex', dbfile): # Add gallery to gallery index table
         print "Added gallery", galleryname, "to galleryindex table"
         new_table(galleryname, dbfile, 'image_db_schema.sql') # Create empty image gallery table if not existing
             
@@ -417,7 +378,7 @@ if __name__ == '__main__':
     
     parser_delimage = subparsers.add_parser('delimage', help='delete one or more images from gallery')
     parser_delimage.add_argument('galleryname', help='name of gallery from which to delete images (table rows)')
-    parser_delimage.add_argument('images', nargs='*', help='image file name(s) to delete')  # Fix so it can't be none ??? --------------------
+    parser_delimage.add_argument('images', nargs='+', help='image file name(s) to delete')
     
     parser_adduser = subparsers.add_parser('adduser', help='add new user to user table')
     parser_adduser.add_argument('username', help='username of the new user')
@@ -428,7 +389,7 @@ if __name__ == '__main__':
     
     parser_addimages = subparsers.add_parser('addimages', help='add images to image gallery table')
     parser_addimages.add_argument('galleryname', help='name of image gallery table to record images in') # galleryname is also recorded (on a row) in gallery index table
-    parser_addimages.add_argument('images', nargs='*', help='image file(s) to add')
+    parser_addimages.add_argument('images', nargs='+', help='image file(s) to add')
     
     parser_test = subparsers.add_parser('test', help='for testing')
     
@@ -442,9 +403,12 @@ if __name__ == '__main__':
         print_table(args.table, args.dbfile)
         
     if args.subparser_name == 'deltable':
-        if raw_input("Really delete table " + args.table + "? (y/n) ") in 'Yy':
-            if delete_table(args.table, args.dbfile):
-                print "Deleted table", args.table, "in file", args.dbfile
+        if find_table(args.table, args.dbfile):
+            if raw_input("Really delete table " + args.table + "? (y/n) ") in 'Yy':
+                if delete_table(args.table, args.dbfile):
+                    print "Deleted table", args.table, "in file", args.dbfile
+        else:
+            print "Could not find table", args.table, "in", args.dbfile
                 
     if args.subparser_name == 'findtable':
         if find_table(args.table, args.dbfile):
@@ -468,20 +432,20 @@ if __name__ == '__main__':
     # User operations
     
     if args.subparser_name == 'adduser':
-        print "Adding user " + args.username + " ..."
-        pw = raw_input("Please enter user's new password: ")
-        pa = raw_input("Repeat password: ")
-        if pw != pa:
-            print "Passwords don't match"
-            sys.exit(0)
+        pw = unicode(raw_input("Please give " + args.username + " a password: "), 'utf-8')
+        #pa = unicode(raw_input("Repeat password: "), 'utf-8')
+        #if pw != pa:
+        #    print "Passwords don't match"
+        #    sys.exit(0)
         hashed_pw = bcrypt.hashpw(pw.encode('utf-8'), bcrypt.gensalt())
         f = raw_input("User's full name (first and last with space between): ")
         gr = raw_input("The groups user is member of (e.g. 'friends,relatives'): ").lower()
         j = time.asctime(time.localtime(time.time()))
         
+        new_table(args.table, args.dbfile, 'user_db_schema.sql') # In case it doesn't exist
         if add_user(args.username, hashed_pw, f, j, gr, args.table, args.dbfile):
             print "Successfully added user %s to table %s in file %s" \
-            % (args.name, args.table, args.dbfile)
+            % (args.username, args.table, args.dbfile)
 
     if args.subparser_name == 'deleteuser':
         if delete_user(args.username, args.table, args.dbfile):
@@ -492,7 +456,7 @@ if __name__ == '__main__':
     
     if args.subparser_name == 'newgallery':
         desc, tags, zipfile, ur, uw, gr, gw = get_user_input('gallery', args.galleryname)
-        new_gallery(args.galleryname, os.getcwd(), desc, tags, zipfile, ur, uw, gr, gw, args.dbfile)
+        create_gallery(args.galleryname, os.getcwd(), desc, tags, zipfile, ur, uw, gr, gw, args.dbfile)
         
     if args.subparser_name == 'delgallery':
         if args.galleryname == 'galleryindex':
@@ -503,14 +467,14 @@ if __name__ == '__main__':
                     print "Deleted image gallery table '" + args.galleryname + "'"
             else:
                 print "Could not find image gallery table '" + args.galleryname + "'"
-            if get_column('gallery_name', args.galleryname, 'gallery_id', 'galleryindex', args.dbfile): # Then the row in galleryindex
+            if get_column('gallery_name', args.galleryname, 'id', 'galleryindex', args.dbfile): # Then the row in galleryindex
                 if delete_row('gallery_name', args.galleryname, 'galleryindex', args.dbfile) > 0:
                     print "Deleted row '" + args.galleryname + "' in galleryindex table"
             else:
                 print "Gallery '" + args.galleryname + "' not found in galleryindex table."
 
     if args.subparser_name == 'delimage':
-        if raw_input('Really delete image(s)? (y/n): ') in 'Yy':
+        if raw_input('Proceed with delete image(s)? (y/n): ') in 'Yy':
             for img in args.images:
                 if delete_row('imagefile', img, args.galleryname, args.dbfile):
                     print "Deleted image", img, "from gallery", args.galleryname
@@ -519,15 +483,23 @@ if __name__ == '__main__':
         if not find_table(args.galleryname, args.dbfile): # Gallery doesn't exist yet
             print "Creating gallery", args.galleryname
             desc, tags, zipfile, ur, uw, gr, gw = get_user_input('gallery', args.galleryname)
-            defpath = get_defpath(os.path.dirname(os.path.realpath(args.images[0]))) # Directory of first image is default defpath
-            new_gallery(args.galleryname, defpath, desc, tags, zipfile, ur, uw, gr, gw, args.dbfile)
+            gpath = get_gpath(os.path.dirname(os.path.realpath(args.images[0]))) # Directory of first image is default gpath
+            create_gallery(args.galleryname, gpath, desc, tags, zipfile, ur, uw, gr, gw, args.dbfile)
         else:
             print "Found gallery", args.galleryname
-            desc, tags, zipfile, ur, uw, gr, gw = get_user_input('images', args.galleryname) # (zipfile not used on this row)
-            zipfile = get_column('galleryname', args.galleryname, 'zipfile', 'galleryindex', args.dbfile)
-            defpath = os.path.normpath(get_column('galleryname', args.galleryname, 'defpath', 'galleryindex', args.dbfile))
+            #desc, tags, zipfile, ur, uw, gr, gw = get_user_input('images', args.galleryname) # (zipfile not used on this row)
+            desc = get_column('gallery_name', args.galleryname, 'description', 'galleryindex', args.dbfile)[0]
+            tags = get_column('gallery_name', args.galleryname, 'tags', 'galleryindex', args.dbfile)[0]
+            ur = get_column('gallery_name', args.galleryname, 'users_r', 'galleryindex', args.dbfile)[0]
+            uw = get_column('gallery_name', args.galleryname, 'users_w', 'galleryindex', args.dbfile)[0]
+            gr = get_column('gallery_name', args.galleryname, 'groups_r', 'galleryindex', args.dbfile)[0]
+            gw = get_column('gallery_name', args.galleryname, 'groups_w', 'galleryindex', args.dbfile)[0]
+            zipfile = get_column('gallery_name', args.galleryname, 'zipfile', 'galleryindex', args.dbfile)[0]
+            gpath = os.path.normpath(get_column('gallery_name', args.galleryname, 'gpath', 'galleryindex', args.dbfile)[0])
         add_images(args.images, desc, tags, ur, uw, gr, gw, args.galleryname, args.dbfile) # Finally, add images
-        add_to_zip(zipfile, defpath, args.images)
+        add_to_zip(zipfile, gpath, args.images)
         
     if args.subparser_name == 'test':
-        print get_column('gallery_name', 'runor', 'gallery_id', 'galleryindex', args.dbfile)
+        print get_column('gallery_name', 'runor', 'id', 'galleryindex', args.dbfile)
+
+# https://www.joelonsoftware.com/2003/10/08/the-absolute-minimum-every-software-developer-absolutely-positively-must-know-about-unicode-and-character-sets-no-excuses/
